backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elevenlabs>=1.57.0",
    "fastapi>=0.115.12",
    "openai>=1.77.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "websockets>=15.0.1",
]
